% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcvusMAR.R
\name{print.vus}
\alias{print.vus}
\title{Print summary results of VUS}
\usage{
\method{print}{vus}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "vus", a result of a call to \code{\link{vus}}.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.vus} prints the results for the output of function \code{\link{vus}}.
}
\details{
\code{print.vus} shows a nice format of the summary table for the VUS estimate results. Some information on the diagnostic test, the fitted values of VUS, and confidence intervals are shown.
}
\seealso{
\code{\link{vus}}
}
