% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation-KNN.R
\name{CVknn}
\alias{CVknn}
\title{Cross-validation for K nearest-neighbor regression}
\usage{
CVknn(X, Dvec, V, k.list = NULL, type, plot = FALSE)
}
\arguments{
\item{X}{a numeric design matrix, which used in \code{\link{rhoKNN}} to estimate probabilities of the disease status.}

\item{Dvec}{a n * 3  binary matrix with three columns, corresponding to the three classes of the disease status. In row i, 1 indicates, the i-th subject belongs to class j, with j = 1, 2, 3. A row of \code{NA} values indicates a non-verified subject.}

\item{V}{a binary vector containing the verification status (1 verified, 0 not verified).}

\item{k.list}{a list of candidate values for finding the optimal value of K. If \code{NULL}(the default) then a set of \eqn{\{1, 2, ..., n.ver\}} is employed. Here, \eqn{n.ver} is the number of verified subjects.}

\item{type}{a type of distance, see \code{\link{rhoKNN}} for more details.}

\item{plot}{if \code{TRUE}, a plot of cross-validation prediction error is produced.}
}
\value{
A good choice for K is returned.
}
\description{
This function calculates the estimated cross-validation prediction error for K nearest-neighbor regression and returns a good choice for K.
}
\details{
The data are divided into 2 groups, verified and non verified. In the verified group, the discrepancy between the true disease status and the KNN estimates of the probabilities of the disease status is computed by varying \code{k} from 1 to the number of verification subjects, see To Duc et al (2016). The optimal value of \code{k} is the point that corresponds to the smallest value of the discrepancy.
}
\examples{
data(EOC)
XX <- cbind(EOC$CA125, EOC$CA153, EOC$Age)
Dna <- preDATA(EOC$D, EOC$CA125)
Dvec.na <- Dna$Dvec
CVknn(XX, Dvec.na, EOC$V, type = "mahala", plot = TRUE)

}
\references{
To Duc, K., Chiogna, M., Adimari, G. (2016): Nonparametric Estimation of ROC Surfaces Under Verification Bias. \url{https://arxiv.org/abs/1604.04656v1}. Submitted.
}

